LPS-78583 Extract base class.
